zcash_client_backend: Move essential wallet types into the `wallet` module.
